bugfix and clear up
