+++ src/__init__.py
@@ -1,3 +1,3 @@
-from .integrators import MCIntegrator, MCMCIntegrator
+from .integrators import MonteCarlo, MCMC
 from .maps import VegasMap
 from .utils import RAvg, set_seed, get_device
